\name{check8list}
\alias{check8list}
\title{
checks the components of a list
}
\description{
checks that all \code{nana} are names of the components of the list
\code{lili}.
}
\usage{
check8list(lili, nana)
}
\arguments{
\item{lili}{(\code{list}). The list the names of which have to be
checked.}
\item{nana}{(\code{character}). The names to be found.}
}
\value{
Nothing but an error is issued if the check is not positive.
}
\keyword{misc}
