\name{form3etsil}
\alias{form3etsil}
\title{
inverse function of form3liste
}
\description{
returns a vector character of the names from a character string
generated by \code{form3liste}. For the moment, sep cannot be an
empty string.\cr
}
\usage{
form3etsil(cara, none="-.-", OPA="{", CPA="}", opa="(", cpa=")", sep="+")
}
\examples{
 rbsb3k("RESET"); # needed only for R checking, to be forgotten
 uu <- form3liste(letters[1:4]);
 form3etsil(uu);
}
\arguments{
\item{cara}{Character to be considered.}
\item{none}{ idem as in form3liste.}
\item{OPA}{ idem as in form3liste.}
\item{CPA}{ idem as in form3liste.}
\item{opa}{ idem as in form3liste.}
\item{cpa}{ idem as in form3liste.}
\item{sep}{ idem as in form3liste.}
}
\details{
Of course, it is implicitely supposed that the inversion is
unambiguous.\cr The syntax (consistent use of the parentheses and
separators) is checked: an error is issued when not correct.\cr
Important: the possible \\n added by \code{form3liste} is not taken
into account and must be removed before calling this function.
}
\value{
A character vector
}
\keyword{print}
