\name{help8ltype}
\alias{help8ltype}
\title{
(ba) provides information about type of links
}
\description{
This function displays requirement and default values for the
different types of links (ltype) implemented into rebastaba
}
\usage{
help8ltype(quelles="ls")
}
\arguments{
\item{quelles}{(="ls") "all": all ltype information\cr "ls" : list of
available ltype with comment\cr "list": list of availables ltype with
comment\cr "argu": possible arguments for each ltype\cr else the
ltype ones want to get information about; e.g. "normal"}
}
\value{
nothing but a printing is issued
}
