\name{length8ion}
\alias{length8ion}
\title{
returns the length of a 'ion' object
}
\description{
(ba) provides the length of a \code{/ion/}.
}
\usage{
length8ion(x)
}
\examples{
 rbsb3k("RESET"); # needed only for R checking, to be forgotten
 length(rbsb.ion0);
}
\arguments{
\item{x}{the 'ion' object to be measured.}
}
\value{
The number of items in 'x'
}
\keyword{misc}
