\name{length8nom}
\alias{length8nom}
\title{
(ba) returns the length of a 'nom' object
}
\description{
provides the number of nodes of a \code{/nom/}.
}
\usage{
length8nom(x)
}
\examples{
 rsba3k("RESET"); # needed only for R checking, to be forgotten
 length(rbsb.nom1);
}
\arguments{
\item{x}{the 'nom' object to be measured.}
}
\value{
The number of nodes in 'x'
}
\keyword{misc}
