\name{print8arc}
\alias{print8arc}
\title{
(ba) prints an /arc/
}
\description{
prints the matrix defining arcs
}
\usage{
print8arc(x, ...)
}
\examples{
 rsba3k("RESET"); # needed only for R checking, to be forgotten
 print(rbsb.arc0);
}
\arguments{
\item{x}{arc object}
\item{\dots}{Further arguments to be passed to the print function.}
}
\value{
nothing a printing is issued
}
\keyword{print}
