\name{print8win}
\alias{print8win}
\title{
(ba) prints the node-variable names
}
\description{
prints the node-variable names of x (a 'nom' object)
}
\usage{
print8win(x, ..., fo=1)
}
\examples{
 rsba3k("RESET"); # needed only for R checking, to be forgotten
 print(rbsb.win0);
}
\arguments{
\item{x}{nom object}
\item{\dots}{Further arguments to be passed to the print function.}
\item{fo}{(=1) format to use, 1 for developed, 0 for one line.}
}
\value{
nothing a printing is issued
}
\keyword{print}
