\name{rebastaba}
\alias{rebastaba}
\title{
(mi) draws some Bayesian networks for illustration
}
\description{
This function draws some Bayesian networks using rebastaba spelling
to serve as a cover figure for the manual
}
\usage{
rebastaba()
}
\examples{
 rsgn3k("RESET");
 rebastaba();
}
\arguments{
}
\value{
nothing but a plot is performed
}
