\name{valid8alk}
\alias{valid8alk}
\title{
(ba) checks an /alk/
}
\description{
This function checks /alk/ objects in a very light way. For a more
complete checking one must use \code{check8alk}
}
\usage{
valid8alk(object)
}
\arguments{
\item{object}{The alk object to be validated.}
}
\details{
It is the validity method for /alk/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
