\name{valid8daf}
\alias{valid8daf}
\title{
(00) checks a /daf/
}
\description{
This function checks /daf/ objects
}
\usage{
valid8daf(object)
}
\arguments{
\item{object}{The daf object to be validated.}
}
\details{
It is the validity method for /daf/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
