\name{valid8des}
\alias{valid8des}
\title{
checks a /des/
}
\description{
This function checks a /des/ objects
}
\usage{
valid8des(object)
}
\arguments{
\item{object}{The des object to be validated.}
}
\details{
It is the validity method for /des/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
