\name{valid8faux}
\alias{valid8faux}
\title{
checks a /faux/
}
\description{
This function checks a /faux/ objects
}
\usage{
valid8faux(object)
}
\arguments{
\item{object}{The faux object to validate.}
}
\details{
It is the validity method for /faux/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{error}
