\name{valid8ion}
\alias{valid8ion}
\title{
checks a /ion/
}
\description{
(ba) This function checks /ion/ objects
}
\usage{
valid8ion(object)
}
\arguments{
\item{object}{The ion object to be validated.}
}
\details{
It is the validity method for /ion/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
