\name{valid8nom}
\alias{valid8nom}
\title{
checks a /nom/
}
\description{
This function checks /nom/ objects
}
\usage{
valid8nom(object)
}
\arguments{
\item{object}{The nom object to be validated.}
}
\details{
It is the validity method for /nom/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
