\name{valid8pam}
\alias{valid8pam}
\title{
(ba) checks a /pam/
}
\description{
This function checks /pam/ objects
}
\usage{
valid8pam(object)
}
\arguments{
\item{object}{The pam object to be validated.}
}
\details{
It is the validity method for /pam/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
