\name{valid8pgr}
\alias{valid8pgr}
\title{
(ba) checks a /pgr/
}
\description{
This function checks /pgr/ objects
}
\usage{
valid8pgr(object)
}
\arguments{
\item{object}{The pgr object to be validated.}
}
\details{
It is the validity method for /pgr/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
