\name{valid8pta}
\alias{valid8pta}
\title{
(ba) checks a /pta/
}
\description{
This function checks /pta/ objects
}
\usage{
valid8pta(object)
}
\arguments{
\item{object}{The pta object to be validated.}
}
\details{
It is the validity method for /pta/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
