\name{valid8win}
\alias{valid8win}
\title{
(ba) checks a /win/
}
\description{
This function checks /win/ objects
}
\usage{
valid8win(object)
}
\arguments{
\item{object}{The win object to be validated.}
}
\details{
It is the validity method for /win/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
