\name{check8list}
\alias{check8list}
\title{
checks the components of a list
}
\description{
checks that all \code{nana} are components of the list \code{lili}.
}
\usage{
check8list(lili, nana)
}
\arguments{
\item{lili}{(\code{list}). The list to be checked.}
\item{nana}{(\code{character}). The components to be found.}
}
\value{
Nothing but an error is issued if the check is not positive.
}
\keyword{misc}
