\name{close8graph}
\alias{close8graph}
\title{
(00) closes the file open by open8graph
}
\description{
According to the global constant rbsb.mfi close the file open by
open8graph. Also if \code{!rbsb.mba} a pause is issued.
}
\usage{
close8graph(message=rbsb.cha0)
}
\arguments{
\item{message}{ Message to display on the terminal. When empty, no
message will be displayed.}
}
\value{
nothing but the actions indicated in the description field are
performed
}
\keyword{IO}
