\name{print8nom}
\alias{print8nom}
\title{
(ba) prints the node-variable names
}
\description{
prints the node-variable names of x (a 'nom' object)
}
\usage{
print8nom(x, ..., what="v", type=0)
}
\examples{
 rsba3k("RESET"); # needed only for R checking, to be forgotten
 print(rbsb.nom2);
}
\arguments{
\item{x}{nom object}
\item{\dots}{Further arguments to be passed to the print function.}
\item{what}{(='v') what to print ('v'=node-variable names; 'n' only
no}
\item{type}{(=0) type of printing (0) a node a line; (1) everything
on the same line.}
}
\value{
nothing a printing is issued
}
\keyword{print}
